refactor: 모달을 하나의 컴포넌트에서 다루던 부분을 모달 공통 레이아웃을 가지는 컴포넌트로 수정
